Split test_excel into submodule
